Ganesh main1

+++ src/main/java/com/spring/jwt/repository/ExpectationsCompleteRepository.java
@@ -3,7 +3,7 @@ package com.spring.jwt.repository;
 import com.spring.jwt.entity.ExpectationsComplete;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExpectationsCompleteRepository extends JpaRepository<ExpectationsComplete, Long> {
+public interface ExpectationsCompleteRepository extends JpaRepository<ExpectationsComplete, Integer> {
     ExpectationsComplete findByUserId(Integer userId);
 
 }
